requirements.txt
streamlit==1.29.0
gradio==4.8.0
python-docx==1.1.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.12
chromadb==0.4.18
sentence-transformers==2.2.2
openai==1.10.0
google-generativeai>=0.3.0
pydantic==2.5.2
pandas==2.1.4
numpy==1.26.4
python-dotenv==1.0.0
typing-extensions==4.9.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
